requirements.txt
Pillow
numpy
gradio
openai